Consolidate auth methods
