+++ lib/auth/logout.ts
